+++ src/components/Modal.tsx
@@ -3,15 +3,13 @@ import { useEffect, useRef } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { arrowLeft } from '@wordpress/icons';
 import { Dialog } from '@headlessui/react';
-import classNames from 'classnames';
 import { AnimatePresence, motion } from 'framer-motion';
+import { ModalCloseButton } from '../layouts/ModalCloseButton';
 import { models } from '../models';
 import { StableDiffusion } from '../models/StableDiffusion';
-import { useAuthStore, useAuthStoreReady } from '../state/auth';
+import { useAuthStore } from '../state/auth';
 import { useGlobalState } from '../state/global';
 import { AvailableModels, ImageLike } from '../types';
-import { Login, LoginWrapper } from './Login';
-import { ModalCloseButton } from './ModalCloseButton';
 
 type ModalProps = {
     setImage: (image: ImageLike) => void;
@@ -33,10 +31,13 @@ export const Modal = ({ setImage, onClose }: ModalProps) => {
             className="stable-diffusion-editor stable-diffusion-modal"
             initialFocus={initialFocus}
             key="main-modal"
-            open={Boolean(currentInterface)}
+            open={Boolean(currentInterface) && Boolean(apiToken)}
             onClose={onClose}>
             <div className="absolute mx-auto w-full h-full overflow-hidden md:p-8 md:flex justify-center items-center z-high">
-                <div className="fixed inset-0 bg-black/60" aria-hidden="true" />
+                <div
+                    className="fixed inset-0 bg-black/60 backdrop-blur-sm"
+                    aria-hidden="true"
+                />
                 <AnimatePresence>
                     <motion.div
                         key="modal"
@@ -44,11 +45,7 @@ export const Modal = ({ setImage, onClose }: ModalProps) => {
                         initial={{ y: 5 }}
                         animate={{ y: 0 }}
                         exit={{ y: 0, opacity: 0 }}
-                        className={classNames(
-                            'sm:flex relative shadow-2xl sm:overflow-hidden max-w-screen-2xl mx-auto bg-white',
-                            // Full height on model interfaces (kind of a hack to keep the login screen constrained)
-                            { 'h-full': apiToken },
-                        )}>
+                        className="sm:flex relative shadow-2xl sm:overflow-hidden max-w-screen-2xl mx-auto bg-white h-full">
                         <Dialog.Title className="sr-only">
                             {
                                 models.find((m) => m.id === currentInterface)
@@ -89,26 +86,6 @@ const ModalContent = ({
     initialFocus,
     onGoBack,
 }: ModalContent) => {
-    const { apiToken } = useAuthStore();
-    const ready = useAuthStoreReady();
-    if (!ready) return null;
-
-    if (!apiToken) {
-        const title = __(
-            'Log in to your Replicate account',
-            'stable-diffusion',
-        );
-        return (
-            <LoginWrapper>
-                <ContentWrapper
-                    onClose={onClose}
-                    title={title}
-                    onGoBack={onGoBack}>
-                    <Login initialFocus={initialFocus} />
-                </ContentWrapper>
-            </LoginWrapper>
-        );
-    }
     if (modelName === 'stability-ai/stable-diffusion') {
         return (
             <ContentWrapper
@@ -137,7 +114,6 @@ const ContentWrapper = ({
     onClose,
     onGoBack,
 }: ContentWrapperProps) => {
-    const { apiToken } = useAuthStore();
     return (
         <>
             <div className="flex items-center justify-between w-full border-b p-4 gap-x-4 fixed md:static top-0 bg-white">
@@ -147,12 +123,7 @@ const ContentWrapper = ({
                 </div>
                 <ModalCloseButton onClose={onClose} />
             </div>
-            <div
-                className={classNames(
-                    'overflow-y-scroll md:flex flex-grow w-screen max-w-full md:pt-0 bg-gray-50 divide-x',
-                    // Full height on model interfaces (kind of a hack to keep the login screen constrained)
-                    { 'h-screen lg:h-auto pt-16': apiToken },
-                )}>
+            <div className="overflow-y-scroll md:flex flex-grow w-screen max-w-full md:pt-0 bg-gray-50 divide-x h-screen lg:h-auto pt-16">
                 {children}
             </div>
         </>
